Fold the Blake3 arch kernels into a blake3 submodule
Follow-up to the review on #2353.

The three Blake3 files sat side by side at the crate root, with the
arch-specific kernels reachable only through `#[path]` attributes on
`blake3_portable`. Turn `blake3` into a real module directory instead,
so the file tree matches the module tree.

Before:

    src/blake3.rs             pub mod blake3
    src/blake3_portable.rs    pub mod blake3_portable
    src/blake3_neon.rs        #[path] child of blake3_portable
    src/blake3_avx512.rs      #[path] child of blake3_portable

After:

    src/blake3/mod.rs         pub mod blake3
    src/blake3/portable.rs    pub mod blake3::portable
    src/blake3/neon.rs        private mod blake3::neon
    src/blake3/avx512.rs      private mod blake3::avx512

The Blake3 spec constants move up into `blake3/mod.rs`, `MSG_SCHEDULE`
with them. They are facts about the hash, not about the portable kernel,
and both `portable` and `neon` read them. A private item in a parent is
visible to its descendants, so every kernel reaches them as `super::IV`
with no visibility annotations.

Pure reorganization: no logic, no lane counts, and no dispatch
conditions change.
