+++ crates/hash/src/blake3/neon.rs
